+++ Reqnroll/Events/ExecutionEvent.cs
@@ -1,10 +1,14 @@
 using System;
 using Reqnroll.Bindings;
+using Reqnroll.Infrastructure;
 
 namespace Reqnroll.Events
 {
     public class ExecutionEvent : IExecutionEvent
     {
+        public DateTime Timestamp { get; }
+
+        public ExecutionEvent() => Timestamp = DateTime.UtcNow;
     }
 
     public class TestRunStartedEvent : ExecutionEvent
@@ -167,10 +171,17 @@ namespace Reqnroll.Events
     public class HookBindingStartedEvent : ExecutionEvent
     {
         public IHookBinding HookBinding { get; }
+        public IContextManager ContextManager { get; private set; }
 
         public HookBindingStartedEvent(IHookBinding hookBinding)
         {
             HookBinding = hookBinding;
+        }
+
+        public HookBindingStartedEvent(IHookBinding hookBinding, IContextManager contextManager) 
+        {
+            HookBinding = hookBinding;
+            ContextManager = contextManager;
         }
     }
 
@@ -179,11 +190,21 @@ namespace Reqnroll.Events
         public IHookBinding HookBinding { get; }
 
         public TimeSpan Duration { get; }
+        public IContextManager ContextManager { get; private set; }
+        public Exception HookException { get; private set; }
 
         public HookBindingFinishedEvent(IHookBinding hookBinding, TimeSpan duration)
         {
             HookBinding = hookBinding;
             Duration = duration;
+        }
+
+        public HookBindingFinishedEvent(IHookBinding hookBinding, TimeSpan duration, IContextManager contextManager, Exception hookException = null) 
+        {
+            HookBinding = hookBinding;
+            Duration = duration;
+            ContextManager = contextManager;
+            HookException = hookException;
         }
     }
 
@@ -193,20 +214,38 @@ namespace Reqnroll.Events
     public class OutputAddedEvent : ExecutionEvent, IExecutionOutputEvent
     {
         public string Text { get; }
+        public FeatureInfo FeatureInfo { get; }
+        public ScenarioInfo ScenarioInfo { get; }
 
         public OutputAddedEvent(string text)
         {
             Text = text;
+        }
+
+        public OutputAddedEvent(string text, FeatureInfo featureInfo, ScenarioInfo scenarioInfo)
+        {
+            Text = text;
+            FeatureInfo = featureInfo;
+            ScenarioInfo = scenarioInfo;
         }
     }
 
     public class AttachmentAddedEvent : ExecutionEvent, IExecutionOutputEvent
     {
         public string FilePath { get; }
+        public FeatureInfo FeatureInfo { get; }
+        public ScenarioInfo ScenarioInfo { get; }
 
         public AttachmentAddedEvent(string filePath)
         {
             FilePath = filePath;
+        }
+
+        public AttachmentAddedEvent(string filePath, FeatureInfo featureInfo, ScenarioInfo scenarioInfo)
+        {
+            FilePath = filePath;
+            FeatureInfo = featureInfo;
+            ScenarioInfo = scenarioInfo;
         }
     }
 }