+++ Reqnroll/Infrastructure/TestExecutionEngine.cs
@@ -103,25 +103,31 @@ namespace Reqnroll.Infrastructure
 
             if (_analyticsTransmitter.IsEnabled)
             {
-                try
-                {
-                    var testAssemblyName = _testRunnerManager.TestAssembly.GetName().Name;
-                    var projectRunningEvent = _analyticsEventProvider.CreateProjectRunningEvent(testAssemblyName);
-                    await _analyticsTransmitter.TransmitReqnrollProjectRunningEventAsync(projectRunningEvent);
-                }
-                catch (Exception)
-                {
-                    // catch all exceptions since we do not want to break anything
-                }
+                _ = Task.Run(TryTransmitReqnrollProjectRunningEventAsync);
             }
 
             _testRunnerStartExecuted = true;
 
-            _testThreadExecutionEventPublisher.PublishEvent(new TestRunStartedEvent());
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new TestRunStartedEvent());
 
             // The 'FireEventsAsync' call might throw an exception if the related before test run hook fails,
             // but we can let the exception propagate to the caller.
             await FireEventsAsync(HookType.BeforeTestRun);
+        }
+
+        async Task TryTransmitReqnrollProjectRunningEventAsync()
+        {
+            try
+            {
+                var testAssemblyName = _testRunnerManager.TestAssembly.GetName().Name;
+                var projectRunningEvent = _analyticsEventProvider.CreateProjectRunningEvent(testAssemblyName);
+                await _analyticsTransmitter.TransmitReqnrollProjectRunningEventAsync(projectRunningEvent);
+            }
+            catch (Exception ex)
+            {
+                // catch all exceptions since we do not want to break anything
+                Debug.WriteLine(ex, "Sending telemetry failed");
+            }
         }
 
         public virtual async Task OnTestRunEndAsync()
@@ -142,7 +148,7 @@ namespace Reqnroll.Infrastructure
             }
             finally
             {
-                _testThreadExecutionEventPublisher.PublishEvent(new TestRunFinishedEvent());
+                await _testThreadExecutionEventPublisher.PublishEventAsync(new TestRunFinishedEvent());
             }
         }
 
@@ -150,7 +156,7 @@ namespace Reqnroll.Infrastructure
         {
             _contextManager.InitializeFeatureContext(featureInfo);
 
-            _testThreadExecutionEventPublisher.PublishEvent(new FeatureStartedEvent(FeatureContext));
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new FeatureStartedEvent(FeatureContext));
 
             // The 'FireEventsAsync' call might throw an exception if the related before feature hook fails,
             // but we can let the exception propagate to the caller.
@@ -181,7 +187,7 @@ namespace Reqnroll.Infrastructure
                     _testTracer.TraceDuration(duration, "Feature: " + FeatureContext.FeatureInfo.Title);
                 }
 
-                _testThreadExecutionEventPublisher.PublishEvent(new FeatureFinishedEvent(FeatureContext));
+                await _testThreadExecutionEventPublisher.PublishEventAsync(new FeatureFinishedEvent(FeatureContext));
 
                 _contextManager.CleanupFeatureContext();
             }
@@ -194,7 +200,7 @@ namespace Reqnroll.Infrastructure
 
         public virtual async Task OnScenarioStartAsync()
         {
-            _testThreadExecutionEventPublisher.PublishEvent(new ScenarioStartedEvent(FeatureContext, ScenarioContext));
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new ScenarioStartedEvent(FeatureContext, ScenarioContext));
 
             try
             {
@@ -263,19 +269,20 @@ namespace Reqnroll.Infrastructure
             }
             finally
             {
-                _testThreadExecutionEventPublisher.PublishEvent(new ScenarioFinishedEvent(FeatureContext, ScenarioContext));
+                await _testThreadExecutionEventPublisher.PublishEventAsync(new ScenarioFinishedEvent(FeatureContext, ScenarioContext));
+
                 _contextManager.CleanupScenarioContext();
             }
         }
 
-        public virtual void OnScenarioSkipped()
+        public virtual async Task OnScenarioSkippedAsync()
         {
             // after discussing the placement of message sending points, this placement causes far less effort than rewriting the whole logic
             _contextManager.ScenarioContext.ScenarioExecutionStatus = ScenarioExecutionStatus.Skipped;
 
             // in case of skipping a Scenario, the OnScenarioStart() is not called, so publish the event here
-            _testThreadExecutionEventPublisher.PublishEvent(new ScenarioStartedEvent(FeatureContext, ScenarioContext));
-            _testThreadExecutionEventPublisher.PublishEvent(new ScenarioSkippedEvent());
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new ScenarioStartedEvent(FeatureContext, ScenarioContext));
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new ScenarioSkippedEvent());
         }
 
         public virtual void Pending()
@@ -309,11 +316,11 @@ namespace Reqnroll.Infrastructure
             await FireScenarioEventsAsync(HookType.AfterStep);
         }
         
-        protected virtual void OnSkipStep()
+        protected virtual async Task OnSkipStepAsync()
         {
             _contextManager.StepContext.Status = ScenarioExecutionStatus.Skipped;
             _testTracer.TraceStepSkipped();
-            _testThreadExecutionEventPublisher.PublishEvent(new StepSkippedEvent());
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new StepSkippedEvent());
 
 
             var skippedStepHandlers = _contextManager.ScenarioContext.ScenarioContainer.ResolveAll<ISkippedStepHandler>().ToArray();
@@ -332,7 +339,7 @@ namespace Reqnroll.Infrastructure
 
         private async Task FireEventsAsync(HookType hookType)
         {
-            _testThreadExecutionEventPublisher.PublishEvent(new HookStartedEvent(hookType, FeatureContext, ScenarioContext, _contextManager.StepContext));
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new HookStartedEvent(hookType, FeatureContext, ScenarioContext, _contextManager.StepContext));
             var stepContext = _contextManager.GetStepContext();
 
             var matchingHooks = _bindingRegistry.GetHooks(hookType)
@@ -364,7 +371,7 @@ namespace Reqnroll.Infrastructure
             //A plugin-hook should not throw an exception under normal circumstances, exceptions are not handled/caught here
             FireRuntimePluginTestExecutionLifecycleEvents(hookType);
 
-            _testThreadExecutionEventPublisher.PublishEvent(new HookFinishedEvent(hookType, FeatureContext, ScenarioContext, _contextManager.StepContext, hookException));
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new HookFinishedEvent(hookType, FeatureContext, ScenarioContext, _contextManager.StepContext, hookException));
 
             //Note: the (user-)hook exception (if any) will be thrown after the plugin hooks executed to fail the test with the right error
             if (hookException != null) ExceptionDispatchInfo.Capture(hookException).Throw();
@@ -382,16 +389,23 @@ namespace Reqnroll.Infrastructure
             var currentContainer = GetHookContainer(hookType);
             var arguments = ResolveArguments(hookBinding, currentContainer);
 
-            _testThreadExecutionEventPublisher.PublishEvent(new HookBindingStartedEvent(hookBinding));
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new HookBindingStartedEvent(hookBinding, _contextManager));
             var durationHolder = new DurationHolder();
-
+            Exception exceptionthrown = null;
             try
             {
                 await invoker.InvokeBindingAsync(hookBinding, _contextManager, arguments, _testTracer, durationHolder);
             }
+            catch (Exception exception)
+            {
+                // This exception is caught in order to be able to inform consumers of the HookBindingFinishedEvent;
+                // The throw; statement ensures that the exception is propagated up to the FireEventsAsync method
+                exceptionthrown = exception;
+                throw;
+            }
             finally
             {
-                _testThreadExecutionEventPublisher.PublishEvent(new HookBindingFinishedEvent(hookBinding, durationHolder.Duration));
+                await _testThreadExecutionEventPublisher.PublishEventAsync(new HookBindingFinishedEvent(hookBinding, durationHolder.Duration, _contextManager, exceptionthrown));
             }
         }
 
@@ -486,7 +500,7 @@ namespace Reqnroll.Infrastructure
 
                 if (isStepSkipped)
                 {
-                    OnSkipStep();
+                    await OnSkipStepAsync();
                 }
                 else
                 {
@@ -586,7 +600,7 @@ namespace Reqnroll.Infrastructure
 
         protected virtual async Task ExecuteStepMatchAsync(BindingMatch match, object[] arguments, DurationHolder durationHolder)
         {
-            _testThreadExecutionEventPublisher.PublishEvent(new StepBindingStartedEvent(match.StepBinding));
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new StepBindingStartedEvent(match.StepBinding));
             
             try
             {
@@ -594,7 +608,7 @@ namespace Reqnroll.Infrastructure
             }
             finally
             {
-                _testThreadExecutionEventPublisher.PublishEvent(new StepBindingFinishedEvent(match.StepBinding, durationHolder.Duration));
+                await _testThreadExecutionEventPublisher.PublishEventAsync(new StepBindingFinishedEvent(match.StepBinding, durationHolder.Duration));
             }
         }
 
@@ -665,7 +679,8 @@ namespace Reqnroll.Infrastructure
                 ? GetCurrentBindingType()
                 : (StepDefinitionType) stepDefinitionKeyword;
             _contextManager.InitializeStepContext(new StepInfo(stepDefinitionType, text, tableArg, multilineTextArg));
-            _testThreadExecutionEventPublisher.PublishEvent(new StepStartedEvent(FeatureContext, ScenarioContext, _contextManager.StepContext));
+
+            await _testThreadExecutionEventPublisher.PublishEventAsync(new StepStartedEvent(FeatureContext, ScenarioContext, _contextManager.StepContext));
 
             try
             {
@@ -674,7 +689,7 @@ namespace Reqnroll.Infrastructure
             }
             finally
             {
-                _testThreadExecutionEventPublisher.PublishEvent(new StepFinishedEvent(FeatureContext, ScenarioContext, _contextManager.StepContext));
+                await _testThreadExecutionEventPublisher.PublishEventAsync(new StepFinishedEvent(FeatureContext, ScenarioContext, _contextManager.StepContext));
                 _contextManager.CleanupStepContext();
             }
         }